moved user profile to redux store. added firestore, shear test form. moved shear calculator components.
