+++ src/components/BSnavbar.js
@@ -6,8 +6,10 @@ import { signIn, signOut } from '../actions/authActions'
 
 class BSnavbar extends Component {
     render() {
-      const { auth } = this.props;
-      console.log(auth);
+      const loginLogoutLinks = this.props.auth.uid ? 
+        <button className="btn btn-outline-primary my-2 my-sm-0" type="submit" onClick={this.props.signOut}>Logout</button>  :
+        <button className="btn btn-outline-primary my-2 my-sm-0" type="submit" onClick={this.props.signIn}>Login</button>
+
       return (
               <div className="navbar navbar-expand-md navbar-dark bg-dark fixed-top row">
                   <a className="navbar-brand" href="/"><img src={Logo} alt="otc-logo"/></a>
@@ -30,9 +32,16 @@ class BSnavbar extends Component {
                             <NavLink className="dropdown-item" to="/tools/accumcalc" activeClassName="active">Accumulator Calculator</NavLink>
                           </div>
                         </li>
+                        <li className="nav-item dropdown">
+                          <NavLink className="nav-link dropdown-toggle" to="/sheartests" id="dropdown01" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false" rel="noreferrer" activeClassName="active">Shear Tests</NavLink>
+                          <div className="dropdown-menu" aria-labelledby="dropdown01">
+                            <NavLink className="dropdown-item" to="/sheartests/add" activeClassName="active">Add Document</NavLink>                    
+                            <NavLink className="dropdown-item" to="/sheartests/search" activeClassName="active">Search</NavLink>
+                          </div>
+                        </li>
                       </ul>
                       <div className=" my-2 my-lg-0">
-                      {auth.user ? <button className="btn btn-outline-primary my-2 my-sm-0" type="submit" onClick={this.props.signOut}>Logout</button> : <button className="btn btn-outline-primary my-2 my-sm-0" type="submit" onClick={this.props.signIn}>Login</button>}
+                      { loginLogoutLinks }
                       </div>
                   </div>
               </div>
@@ -41,9 +50,8 @@ class BSnavbar extends Component {
 };
 
 const mapStateToProps = (state) => {
-    console.log(state);
     return{
-        auth: state.firebase.auth
+      auth: state.firebase.auth
     };
 };
