+++ src/components/ShearCalculator/ShearLinesTable.js
