+++ src/firebase.js
@@ -1,4 +1,5 @@
 import firebase from 'firebase/app'
+import 'firebase/firestore'
 import 'firebase/auth'
 
 const apikey = process.env.REACT_APP_FIREBASE_API;
