+++ src/index.js
@@ -1,14 +1,26 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
-import { createStore, applyMiddleware } from 'redux';
+import { createStore, applyMiddleware, compose } from 'redux';
 import thunk from 'redux-thunk';
+import { reduxFirestore, getFirestore } from 'redux-firestore';
+import { reactReduxFirebase, getFirebase } from 'react-redux-firebase';
+import config from 'firebase';
+
 import App from './App';
 import reducers from './reducers';
 import * as serviceWorker from './serviceWorker';
 
+const store = createStore(reducers,
+    compose(
+        applyMiddleware(thunk.withExtraArgument({getFirebase, getFirestore})),
+        reduxFirestore(config),
+        reactReduxFirebase(config)
+    )
+);
+
 ReactDOM.render(
-    <Provider store={createStore(reducers, applyMiddleware(thunk))}>
+    <Provider store={store}>
         <App />
     </Provider>, 
     document.getElementById('root')
